Sort jdbc drivers (#21)
This fix solves the issue with the JDBC driver ordering in the Preferences/JDBC Management tab. At first, they were ordered alphabetically and that was not desired, so I have ordered them based on the version number.

TableViewerSorter.java now contains a hashmap that can have an arbitrary comparator for each table column (the key of this hashmap is the column value). When sorting a column, I check whether the hashmap contains a comparator for that column or not and, if it exists, use it.