cannot execute a query that it has a sqlmap syntax tags (#41)
* cannot execute a query that it has a sqlmap syntax tags

When the query has sqlmap tags such as <select>, <update>, <insert>,
<delete>, cubrid manager expects to parse as well as transform that
query to a normal query which allows to be executed in the query editor
through parameter binding step.

However, some query might possible to have such tags in conditions. In
this condition, when you run a query, it couldn't execute because
cubrid manager handles that query as a sqlmap rather than a normal
query. So, it is always failed.

For this reason, we should have to provide sqlmap parsing function with
an extra action rather than a query run action. Because even if it has
a function to distinguish between a normal and a sqlmap query in the
query editor, but it is not enough with such simple logic.

In this patch, I added a new menu item in the context menu of the query
editor as `Parse this sqlmap query` which located under the `Run this
query` menu item. When you click this menu, it will parse the query as
the sqlmap and make it open the Sqlmap pane on the right side of the
query editor due to binding parameters.