release 2016.12 (#40)
* master: rewrite some message strings (#22)

* Sort jdbc drivers (#21)

This fix solves the issue with the JDBC driver ordering in the Preferences/JDBC Management tab. At first, they were ordered alphabetically and that was not desired, so I have ordered them based on the version number.

TableViewerSorter.java now contains a hashmap that can have an arbitrary comparator for each table column (the key of this hashmap is the column value). When sorting a column, I check whether the hashmap contains a comparator for that column or not and, if it exists, use it.

* deleted wrong characters in ko_KR messages (#26)

trivial change. so merge directly.

* CloseQueryEditor: if all tabs are closed, start sqlEditorCounter at 1 (#25)

seunghun-kim reviewed, tested and confirmed it operated as expected.

* Repair strings (#24)

* RepairStrings: repair mistakes in labels/buttons

* RepairStrings: update gitignore to ignore metadata

* RepairStrings: revert gitignore

* Scroll drivers list (#23)

* master: create separate class (MessageDialogWithScrollableMessage)

# Conflicts:
#	com.cubrid.common.ui/src/com/cubrid/common/ui/spi/util/CommonUITool.java

* master: add scrollable window with the drivers list

* ScrollDriversList: repair git mistake

* ScrollDriversList: fix another merging mistake

* ScrollDriversList: replace "OK" with Messages.btnOk

* ScrollDriversList: choose whether you show informative message or informative message + scrollable text when opening an information box.

* ScrollDriversList: remove some unused imports

* ScrollDriversList: small improvement

* ScrollDriversList: remove a couple of double lines

* ScrollDriversList: change the same of two variables

* ScrollDriversList: remove redundant setLayout. format code

* ScrollDriversList: delete {0} from jdbcDriverDownloadSuccessMsg.
		   change the way we build the downloadedFiles string

* UIImprovements: -remove superfluous message from the jdbc advanced op… (#27)

* UIImprovements: -remove superfluous message from the jdbc advanced options (load jdbc driver)
		-remove superfluous message that appeared when clicking File->Open SQL File or File->Save all SQL files (Setting SQL file charset).

* RemoveMessages: restore messages

* UIImprovements: transform 1th, 2th and 3th to their correct spelling (#28)

* UIImprovements: transform 1th, 2th and 3th to their correct spelling

* OrdinalNumbers: repair strings for some locales

* OrdinalNumbers: fix typo

* OrdinalNumbers: locale improvements

* OrdinalNumbers: locales are now enums

* OrdinalNumbers: remove unused import

* OrdinalNumbers: refactor for messages in query

* OrdinalNumbers: make the change for task_querydesc

* UIImprovements: -fix the tools->broker log_top to excel labels (they weren't centered) (#29)

* fixed that elapsed time was printed as twice (#34)

* Search button (#31)

* UIImprovements: -fix the "search in cubrid.org" button

* SearchButton: remove the hardcoded stuff

* SearchButton: remove unused imports

* Auto layout button (#33)

- UIImprovements: make the "auto layout" button's height bigger to that the inner text is shown correctly

* AutoLayoutButton: remove hard coded stuff

* AutoLayoutButton: make the font smaller

* AutoLayoutButton: -don't resize the whole toolbar.
		  -don't decrease font's size
		  -put the marginHeight to 0

* AutoLayoutButton: unused import

* FixSearch: add useful keywords to search in preferences (#35)

* ExecuteSQLWindow: fix the query text box (#36)

* Bytes: change from X to XB (X={K, M, G, T}) (#37)

* cannot execute a query that it has a sqlmap syntax tags (#41)

* cannot execute a query that it has a sqlmap syntax tags

When the query has sqlmap tags such as <select>, <update>, <insert>,
<delete>, cubrid manager expects to parse as well as transform that
query to a normal query which allows to be executed in the query editor
through parameter binding step.

However, some query might possible to have such tags in conditions. In
this condition, when you run a query, it couldn't execute because
cubrid manager handles that query as a sqlmap rather than a normal
query. So, it is always failed.

For this reason, we should have to provide sqlmap parsing function with
an extra action rather than a query run action. Because even if it has
a function to distinguish between a normal and a sqlmap query in the
query editor, but it is not enough with such simple logic.

In this patch, I added a new menu item in the context menu of the query
editor as `Parse this sqlmap query` which located under the `Run this
query` menu item. When you click this menu, it will parse the query as
the sqlmap and make it open the Sqlmap pane on the right side of the
query editor due to binding parameters.